Move admin-side observer to separate admin class
